+++ src/Commands/AdminCommands/CleanupCommand.php
@@ -68,7 +68,6 @@ class CleanupCommand extends AdminCommand
      * @var array
      */
     protected static $default_tables_to_clean = [
-        'botan_shortener',
         'callback_query',
         'chosen_inline_result',
         'conversation',
@@ -85,7 +84,6 @@ class CleanupCommand extends AdminCommand
      * @var array
      */
     protected static $default_clean_older_than = [
-        'botan_shortener'      => '30 days',
         'chat'                 => '365 days',
         'callback_query'       => '30 days',
         'chosen_inline_result' => '30 days',
@@ -241,11 +239,6 @@ class CleanupCommand extends AdminCommand
             'conversation'    => ['table' => TB_CONVERSATION, 'field' => 'updated_at'],
             'request_limiter' => ['table' => TB_REQUEST_LIMITER, 'field' => 'created_at'],
         ];
-
-        // Botan table is only available if enabled.
-        if (defined('TB_BOTAN_SHORTENER')) {
-            $simple_tables['botan_shortener'] = ['table' => TB_BOTAN_SHORTENER, 'field' => 'created_at'];
-        }
 
         foreach (array_intersect(array_keys($simple_tables), $tables_to_clean) as $table_to_clean) {
             $queries[] = sprintf(